requirements.txt
# Core
python-telegram-bot>=21.0
python-telegram-bot[job-queue]>=21.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
pgvector>=0.3.0
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.27.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.0.0

# Migrations
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23.0
